pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mindvault-ai"
version = "0.3.2"
description = "Unified knowledge management: Search + Graph + Wiki"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "etinpres" }]

dependencies = [
    "networkx",
    "tree-sitter>=0.23.0",
    "tree-sitter-python",
    "tree-sitter-typescript",
    "tree-sitter-javascript",
    "tree-sitter-go",
    "tree-sitter-rust",
    "tree-sitter-java",
    "tree-sitter-swift",
    "tree-sitter-kotlin",
    "tree-sitter-c",
    "tree-sitter-cpp",
    "tree-sitter-ruby",
    "tree-sitter-c-sharp",
    "python-docx>=1.0",
    "openpyxl>=3.1",
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "build",
    "twine",
]

[project.scripts]
mindvault = "mindvault.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
